♻️ WOW-71 refactor : stores/search-> stores/product로 폴더명 수정
